+++ andes/models/shaft/__init__.py
@@ -1,0 +1,1 @@
+from andes.models.shaft.shaft5 import SHAFT5  # NOQA